backend/requirements.txt
flask>=3.0,<4.0
supabase>=2.0,<3.0
pyjwt>=2.8,<3.0
gunicorn>=21.2,<23.0
hypothesis>=6.0,<7.0
pytest>=8.0,<9.0
flake8>=7.0,<8.0
freezegun>=1.2,<2.0
